+++ index.ts
@@ -4,6 +4,9 @@ import  { uploadToIpfs, store } from "./lib/ipfs"
 import type { Create1155ContractArgs } from "./lib/types/Create1155ContractArgs"
 import use1155Collect from "./hooks/use1155Collect";
 import useZoraFixedPriceSaleStrategy from "./hooks/useZoraFixedPriceSaleStrategy";
+import getEncodedMinterArgs from "./lib/zora/getEncodedMinterArgs";
+import useUniversalMinter from "./hooks/useUniversalMinter";
+import useCollection from "./hooks/useCollection";
 
 export {
     // IPFS 
@@ -11,10 +14,13 @@ export {
     uploadToIpfs, 
 
     // ZORA
-    useCreate1155Contract,
     type Create1155ContractArgs,
+    getEncodedMinterArgs,
     use1155Collect,
+    useCollection,
+    useCreate1155Contract,
     useZoraFixedPriceSaleStrategy,
+    useUniversalMinter,
 
     // ETHERS
     useEthersSigner